requirements.txt
asgiref==3.7.2
Django==5.0
django-environ==0.11.2
djangorestframework==3.14.0
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
wheel==0.41.2
